improvements for multiple aggregations
fix bug where the up metric was duplicated for multiple aggregation. The up metric now got an aggregation name label that shows if the aggregation could be executed successfully.
parallelize aggregation queries against the MongoDB, introduce the option to limit concurrent queries, to avoid overloading the database with monitoring queries.